Change ints to size_ts

+++ src/loop_optimize.cpp
@@ -2,14 +2,14 @@
 
 namespace simd_optimizations
 {
-    float calculateNormalLoop(const float* arr1, const float* arr2, int size)
+    float calculateNormalLoop(const float* arr1, const float* arr2, size_t size)
     {
         return 0.0;
     }
 
-    float calculateOptimizedLoop(const float* arr1, const float* arr2, int size)
+    float calculateOptimizedLoop(const float* arr1, const float* arr2, size_t size)
     {
         return 0.0;
     }
-    
+
 } //end namespace simd_optimizations
